Migrated to new package namespace updated licenses
