+++ README.md
@@ -1,22 +1,22 @@
-#GWT Driver
+# GWT Driver
 GWT bindings for Selenium WebDriver.
 
-##About
+## About
 * <a href="https://github.com/niloc132/gwt-driver/wiki">Concepts</a>
 
-##Documentation
+## Documentation
 * <a href="https://github.com/niloc132/gwt-driver/wiki">Wiki Home Docs</a>
 * <a href="http://niloc132.github.io/gwt-driver/apidocs/">Javadoc</a>
 
-##Download
+## Download
 * <a href="https://github.com/niloc132/gwt-driver/wiki/Maven-Configuration">Maven Configuration</a>
 
-##Extensions
+## Extensions
 * <a href="https://github.com/niloc132/gxt-driver">GXT Driver</a>
 
-##Samples
+## Samples
 * <a href="https://github.com/niloc132/gwt-driver-sample">GWT & GXT Sample</a>
 
-##Videos
+## Videos
 * <a href="https://www.youtube.com/watch?feature=player_detailpage&v=Ko5by01vw_c#t=1261">GWT-Driver at the Twin Cites GWT User Group</a>
 * <a href="http://www.youtube.com/watch?feature=player_detailpage&v=UrkD8CWboDE#t=1472">GWT-Driver at the April 2013 GWT Community Hangout</a>
