+++ src/main/java/com/vertispan/webdriver/gwt/gwtdriver/GwtDriver.gwt.xml
@@ -17,11 +17,16 @@
     limitations under the License.
 
 -->
-<!doctype html>
-<html>
-<head>
-<title>test</title>
-<script src="simplewidgets/simplewidgets.nocache.js"></script>
-</head>
-<body></body>
-</html>
+<module>
+	<inherits name="com.google.gwt.user.User" />
+	<source path="client" />
+
+	<generate-with class="com.vertispan.webdriver.gwt.gwtdriver.SeExporterGenerator" >
+		<when-type-is class="com.vertispan.webdriver.gwt.gwtdriver.client.SeleniumExporter" />
+	</generate-with>
+
+	<define-configuration-property name="gwtdriver.methods" is-multi-valued="true"/>
+	<extend-configuration-property name="gwtdriver.methods" value="com.vertispan.webdriver.gwt.gwtdriver.client.SeleniumExporter.DefaultExportedMethods"/>
+
+	<entry-point class="com.vertispan.webdriver.gwt.gwtdriver.client.SeleniumExporter" />
+</module>
