+++ src/main/java/com/vertispan/webdriver/gwt/gwtdriver/models/Child.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2013 Colin Alworth
+ * Copyright 2012-2013 Sencha Labs
+ * Copyright 2022 Vertispan LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.vertispan.webdriver.gwt.gwtdriver.models;
+
+import com.google.gwt.user.client.ui.Widget;
+
+/**
+ * Experimental PageObject annotation to try finding specific widgets along with other supported
+ * By statements.
+ */
+public @interface Child {
+	Class<? extends Widget> type() default Widget.class;
+}
