--- src/main/java/org/senchalabs/gwt/gwtdriver/GwtDriver.gwt.xml
@@ -1,32 +1,0 @@
-<!--
-  #%L
-  gwt-driver
-  %%
-  Copyright (C) 2012 - 2013 Sencha Labs
-  %%
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  
-       http://www.apache.org/licenses/LICENSE-2.0
-  
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-  #L%
-  -->
-<module>
-	<inherits name="com.google.gwt.user.User" />
-	<source path="client" />
-
-	<generate-with class="org.senchalabs.gwt.gwtdriver.SeExporterGenerator" >
-		<when-type-is class="org.senchalabs.gwt.gwtdriver.client.SeleniumExporter" />
-	</generate-with>
-
-	<define-configuration-property name="gwtdriver.methods" is-multi-valued="true"/>
-	<extend-configuration-property name="gwtdriver.methods" value="org.senchalabs.gwt.gwtdriver.client.SeleniumExporter.DefaultExportedMethods"/>
-
-	<entry-point class="org.senchalabs.gwt.gwtdriver.client.SeleniumExporter" />
-</module>
