--- src/main/java/org/senchalabs/gwt/gwtdriver/models/Child.java
@@ -1,31 +1,0 @@
-package org.senchalabs.gwt.gwtdriver.models;
-
-/*
- * #%L
- * gwt-driver
- * %%
- * Copyright (C) 2012 - 2013 Sencha Labs
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-
-import com.google.gwt.user.client.ui.Widget;
-
-/**
- * Experimental PageObject annotation to try finding specific widgets along with other supported
- * By statements.
- */
-public @interface Child {
-	Class<? extends Widget> type() default Widget.class;
-}
