--- src/main/java/org/senchalabs/gwt/gwtdriver/models/GwtRootPanel.java
@@ -1,48 +1,0 @@
-package org.senchalabs.gwt.gwtdriver.models;
-
-/*
- * #%L
- * gwt-driver
- * %%
- * Copyright (C) 2012 - 2013 Sencha Labs
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.senchalabs.gwt.gwtdriver.models.GwtWidget.ForWidget;
-
-import com.google.gwt.user.client.ui.RootPanel;
-
-/**
- * Represents a {@link RootPanel}. Has convenience constructors for the body tag and elements with an id.
- *
- */
-@ForWidget(RootPanel.class)
-public class GwtRootPanel extends WidgetContainer {
-
-	public GwtRootPanel(WebDriver driver, WebElement element) {
-		super(driver, element);
-	}
-	public GwtRootPanel(WebDriver driver, String id) {
-		this(driver, driver.findElement(By.id(id)));
-	}
-	public GwtRootPanel(WebDriver driver) {
-		this(driver, driver.findElement(By.tagName("body")));
-	}
-	
-	
-}
