+++ src/test/java/com/vertispan/webdriver/gwt/gwtdriver/models/SimpleWidgets.gwt.xml
@@ -1,3 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <!--
 
     Copyright 2013 Colin Alworth
@@ -17,11 +18,14 @@
     limitations under the License.
 
 -->
-<!doctype html>
-<html>
-<head>
-<title>test</title>
-<script src="simplewidgets/simplewidgets.nocache.js"></script>
-</head>
-<body></body>
-</html>
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.5.0.rc1//EN" "http://google-web-toolkit.googlecode.com/svn/tags/2.5.0.rc1/distro-source/core/src/gwt-module.dtd">
+<module rename-to="simplewidgets">
+	<inherits name="com.google.gwt.user.User" />
+	<inherits name="com.vertispan.webdriver.gwt.gwtdriver.GwtDriver" />
+	
+	<inherits name="com.google.gwt.logging.LoggingDisabled" />
+	
+	<source path="client" />
+	<entry-point class="com.vertispan.webdriver.gwt.gwtdriver.models.client.SimpleWidgetsEP" />
+	<collapse-all-properties />
+</module>
