--- src/test/java/org/senchalabs/gwt/gwtdriver/models/SimpleWidgets.gwt.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  #%L
-  gwt-driver
-  %%
-  Copyright (C) 2012 - 2013 Sencha Labs
-  %%
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  
-       http://www.apache.org/licenses/LICENSE-2.0
-  
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-  #L%
-  -->
-
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.5.0.rc1//EN" "http://google-web-toolkit.googlecode.com/svn/tags/2.5.0.rc1/distro-source/core/src/gwt-module.dtd">
-<module rename-to="simplewidgets">
-	<inherits name="com.google.gwt.user.User" />
-	<inherits name="org.senchalabs.gwt.gwtdriver.GwtDriver" />
-	
-	<inherits name="com.google.gwt.logging.LoggingDisabled" />
-	
-	<source path="client" />
-	<entry-point class="org.senchalabs.gwt.gwtdriver.models.client.SimpleWidgetsEP" />
-	<collapse-all-properties />
-</module>
